requirements.txt
# Web Scraping
scrapy==2.11.2
requests==2.31.0
beautifulsoup4==4.12.3

# PDF Processing
pdfminer.six==20231228
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0

# NLP and Embeddings
sentence-transformers==2.7.0
transformers==4.40.2
torch==2.3.0
tiktoken==0.7.0

# Vector Database
faiss-cpu==1.8.0
numpy==1.26.4

# LLM Integration
llama-cpp-python==0.2.77
langchain==0.2.1
langchain-community==0.2.1

# Development Tools
jupyter==1.0.0
black==24.4.2
pytest==8.2.1
python-dotenv==1.0.1

# API Framework
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.3

# Utilities
tqdm==4.66.4
pandas==2.2.2
click==8.1.7